ENGENHARIA_DE_DADOS_II/requirements.txt
absl-py==0.11.0
adal==1.2.2
alabaster==0.7.12
altgraph==0.17
anaconda-client==1.7.2
anaconda-navigator==1.10.0
anaconda-project==0.8.3
aniso8601==9.0.1
appdirs==1.4.4
applicationinsights==0.11.9
apyori==1.1.2
asn1crypto==1.0.1
astroid==2.3.1
astropy==3.2.1
astunparse==1.6.3
atomicwrites==1.3.0
attrs==19.2.0
autograd==1.3
autograd-gamma==0.4.2
Automat==20.2.0
autopep8==1.5.5
azure-common==1.1.25
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.60.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.1.1
azure-mgmt-resource==8.0.1
azure-mgmt-storage==8.0.0
azureml-automl-core==1.0.85.5
azureml-core==1.0.85.5
azureml-dataprep==1.1.38
azureml-dataprep-native==13.2.0
azureml-pipeline==1.0.85
azureml-pipeline-core==1.0.85.1
azureml-pipeline-steps==1.0.85
azureml-sdk==1.0.85
azureml-telemetry==1.0.85.2
azureml-train==1.0.85
azureml-train-automl-client==1.0.85.4
azureml-train-core==1.0.85
azureml-train-restclients-hyperdrive==1.0.85
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.0
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.0
bokeh==1.3.4
boto==2.49.0
Bottleneck==1.2.1
branca==0.4.2
Brotli==1.0.9
cachelib==0.1.1
cachetools==4.1.0
cassandra-driver==3.25.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
colorlog==5.0.1
comtypes==1.1.7
conda==4.9.2
conda-build==3.21.4
conda-package-handling==1.7.2
conda-verify==3.4.2
constantly==15.1.0
contextlib2==0.6.0
cryptography==2.7
cssselect==1.1.0
cssutils==2.3.0
cx-Freeze==6.6
cx-Logging==3.0
cx-Oracle==7.3.0
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
dask==2.5.2
decorator==4.4.0
defusedxml==0.6.0
distfit==1.4.0
distributed==2.5.2
distro==1.4.0
docker==4.2.0
docutils==0.15.2
dotnetcore2==2.1.13
easydev==0.11.1
emoji==1.2.0
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
feather-format==0.4.1
filelock==3.0.12
firebase==3.0.1
fitter==1.3.0
flake8==3.7.9
flasgger==0.9.5
Flask==1.1.2
Flask-Compress==1.8.0
Flask-Cors==3.0.10
Flask-Ext==0.1
Flask-JWT-Extended==4.2.3
Flask-Login==0.5.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.9
flask-restplus==0.13.0
flask-restx==0.3.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
flatbuffers==1.12
folium==0.12.1
fsspec==0.5.2
fusepy==3.0.1
future==0.18.2
gast==0.3.3
gcloud==0.18.3
Genderize==0.3.1
geomet==0.2.1.post1
gevent==1.4.0
glob2==0.7
glpi-api==0.3.0
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==0.4.15
grpcio==1.32.0
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
httpie==2.4.0
httplib2==0.19.0
hyperlink==21.0.0
idna==2.8
imageio==2.6.0
imagesize==1.1.0
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata==3.7.3
incremental==21.3.0
install==1.3.4
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
JayDeBeApi==1.1.1
jdcal==1.4.1
jedi==0.15.1
jeepney==0.4.3
Jinja2==2.10.3
jmespath==0.9.5
joblib==0.13.2
JPype1==0.6.3
json5==0.8.5
jsonpickle==1.3
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
jwcrypto==0.8
kafka-python==2.0.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
libarchive-c==2.8
lifelines==0.24.8
llvmlite==0.29.0
locket==0.2.0
lockfile==0.12.2
lxml==4.4.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mlxtend==0.17.2
mock==3.0.5
mod-wsgi==4.7.1
more-itertools==7.2.0
MouseInfo==0.1.3
mpmath==1.1.0
msgpack==0.6.1
msrest==0.6.11
msrestazure==0.6.2
multipledispatch==0.6.0
mysqlclient==2.0.3
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
ndg-httpsclient==0.5.1
networkx==2.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
notebook-as-pdf==0.4.0
numba==0.45.1
numexpr==2.7.0
numpy==1.19.5
numpydoc==0.9.1
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
openpyxl==3.0.0
opt-einsum==3.3.0
packaging==19.2
pandas==0.25.1
pandocfilters==1.4.2
parsel==1.6.0
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
pathspec==0.7.0
patsy==0.5.1
pefile==2019.4.18
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pid==3.0.4
Pillow==6.2.0
pkginfo==1.5.0.1
plotly==4.6.0
pluggy==0.13.0
ply==3.11
premailer==3.9.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
Protego==0.1.16
protobuf==3.12.0
psutil==5.6.3
ptyprocess==0.7.0
py==1.8.0
pyaes==1.6.1
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pycosat==0.6.3
pycox==0.2.1
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
PyDispatcher==2.0.5
pyee==8.1.0
PyExcelerate==0.10.0
pyflakes==2.1.1
PyGetWindow==0.0.9
Pygments==2.9.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
PyJWT==2.1.0
pylint==2.4.2
pymongo==3.12.0
PyMsgBox==1.0.9
pymssql==2.1.4
PyMySQL==0.10.1
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
PyPDF2==1.26.0
pyperclip==1.8.1
pypickle==0.1.0
pypiwin32==223
pyppeteer==0.2.5
pyreadline==2.1
PyRect==0.1.4
pyrsistent==0.15.4
pysc==0.1.1
PyScreeze==0.1.26
PySocks==1.7.1
pyTelegramBotAPI==3.7.9
pytest==5.2.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-daemon==2.2.4
python-dateutil==2.8.0
python-decouple==3.4
python-dotenv==0.10.5
python-jwt==3.3.0
PyTweening==1.0.3
pytz==2019.3
PyWavelets==1.0.3
pywhatkit==3.8
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
queuelib==1.5.0
redis==3.5.3
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
rope==0.14.0
rsa==4.0
ruamel-yaml==0.15.46
ruamel.yaml==0.15.89
scikit-image==0.15.0
scikit-learn==0.23.2
scipy==1.7.0
Scrapy==2.4.1
seaborn==0.11.0
SecretStorage==3.1.2
Send2Trash==1.5.0
service==0.6.0
service-identity==18.1.0
setproctitle==1.1.10
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
sklearn-pandas==1.8.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
SpeechRecognition==3.8.1
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.9
sseclient==0.0.27
statsmodels==0.12.2
sympy==1.4
tables==3.5.2
tblib==1.4.0
Telethon==1.21.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.2.0
tensorflow-gpu-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
threadpoolctl==2.1.0
tk==0.1.0
toml==0.10.2
toolz==0.10.0
torch==1.5.0
torchtuples==0.2.0
torchvision==0.6.0
tornado==6.0.3
tqdm==4.57.0
traitlets==4.3.3
Twisted==21.2.0
twisted-iocpsupport==1.0.1
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.25.11
w3lib==1.22.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==0.16.0
wfastcgi==3.0.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wordcloud==1.7.0
wrapt==1.12.1
WTForms==2.2.1
xgboost==1.4.2
xlrd==1.2.0
XlsxWriter==1.2.1
xlwings==0.15.10
xlwt==1.3.0
xmltodict==0.12.0
yagmail==0.14.256
yapf==0.30.0
zict==1.0.0
zipp==0.6.0
zope.interface==5.3.0
